backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Image Processing
opencv-python-headless>=4.10.0
numpy>=1.26.0
pillow>=10.0.0
scikit-image>=0.22.0

# Geospatial
rasterio>=1.3.0
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utils
aiofiles>=23.0.0
uuid6>=2024.1.12
